Make an attempt at verifying the portability claims in the README
One of the explicit goals of this project is to make this implementation
of Forth easily portable across microcontrollers. As a first step, this
is an attempt at porting the existing code from nasm to GNU assembler
which should ideally be an easy first "port" since I'm not even changing
the CPU, just the assembler.

This commit reprises the work done in the very first commit done using
nasm and simply confirms that I can get to a "halt"ing interpreter