Modify genprims to generate GNU "as" friendly prims.s
The current version of genprims generates assembly code which can be
assembled by the nasm assembler. Since this porting exercise targets
the GNU "as" assembler, the genprims script needs to be adapted to
generate code for "as" instead.

The only change that was needed was for comments, since "as" uses '#'
while nasm uses ';'

$TECHDEBT ^ genprims is almost an exact copy of the nasm version