Implement literals
One thing I learnt from the initial bootstrap exercise is that the branch
operators should be implemented as early as possible so that the regression
tests become easy to write. Also since all of the branch generation logic
in genrom is already available, it makes sense to go ahead and leverage it.

Since the branch operators depend on literals, I need to implement it first