Build the GNU "as" port as part of top level builds
With the addition of "rot" in the previous commit, the 256 byte area for
primitives is now full. So I'll consider the GNU "as" port complete since
all of the remaining code which can be added will be architecture
independent Forth code.

Although I've treated this port as secondary to the nasm build, I want to
ensure the code doesn't rot away over time, so I've added the build to the
top level make (as a clunky subdirectory make addon, since I'm being lazy).

Some documentation about the porting details and planning for the future
ports (including immediate next steps) is in PORTING.