C Port : Add error handling, for invalid opcodes
Since the rom code is auto-generated, any opcode errors are expected to
be caught early on. But just to be defensive, I've added some error
handling in case an invalid opcode manages to sneak into the ROM.

This is obviously trivial in C but much harder to pull off in assembly,
so why not use the available powerful capability to speed up debugging
in case there are subtle bugs.