Redo the C port using subroutine threaded code ("STC")
While trying to add branch support I realized that C does not have
a convenient means of generating offsets that is available when using
assemblers although I assume it could be done using the preprocessor or
by extending genrom to generate the offsets directly. For now, I've
decided to instead generate the C code directly (in Forth-lore, this
technique is called "subroutine threaded code"/STC)

Since this effectively requires a rewind from the beginning, I'm doing this
over with a reimplementation of code.prims that only uses "halt" to begin
again. The new "genstc" Perl script is a replacement for the "genrom"
script. Since I expect that most of the code in genrom can be reused, I'll
keep it around for now rather than delete it right away.