C Port : Implement conditionals
Generating C code to do branching makes this an exercise in triviality,
but not too surprisingly, this strategy works quite well even if the
generated code was assembly (or pretty much any other language or even
machine code).

The downside to generating "subroutine threaded code" in this fashion
is that it makes creating the full Forth REPL a separate bootstrap
exercise since the "jmp" and "jz" operators are not available at runtime
and so generating code dynamically for conditionals at runtime is not
possible in this scheme of things.