Use the existing x86/nasm port to verify the new rom.4th tests
Although the new generic/"toplevel" rom.4th was based on the existing
x86/x86-as rom.4th, it makes sense to verify that it hasn't regressed
anything in the current code base before starting to use it in future
ports to run tests on a completely new code base.

With this set of changes, steps 0-15 (out of the max 37) are working as
expected. The only substantial change that was needed was to defs.4th
and that was only because the old echon/getc logic is now part of rom.4th

Rather than turn this into one single extra large commit, I'll do this
piecemeal, fixing each broken test in a separate commit. Currently it
breaks at step 16 so I'll work on fixing up that next.