Fix x86 ports : add code for "stick"
The new version of rom.4th expects a "stick" primitive - this is in addition
to "pick" (and was added for symmetry with it). This change adds the code
for "stick" to code.prims in both x86 (nasm) and x86-as (GNU "as").

I was in two minds about whether to keep the implementation as simple as
possible or to make the stack indexing consistent between "stick" and
"pick". For now, I've decided on keeping the implementation simple so that
the code for "pick" and "stick" look identical.