Change the test input file format to match rom.4th's expectation
The newer version of rom.4th expects the test input to have a certain
character sequence. Previously the entire sequence (in test.inp) was
pumped in as a single string during the test but this doesn't quite
work when the input is read conditionally (for example, the old getc
code reads a character only when something is available on the input port).

To fix this, since the expected output file is processed by the fpp script,
it makes sense to also have the input file also be pre-processed using
the same script.