Ensure that the x86/notest target continues to work
So far, all of the testing has focused on making sure that the tests added
as part of the newer version of rom.4th run successfully. Outside of the
test framework, there is an x86/notest make target and its purpose was to
check the minimal code size required by this implementation (when all the
tests are disabled). That part appears to have suffered some code rot and
is fixed as part of this commit. It turned out to be a bug in the new fpp
script where the state machine to describe the behaviour was incorrect and
so the code that implemented it was also incorrect.