PDP11 Port : Step 0 : A halt'ing interpreter
Since the pdp11 port is starting from scratch, I've added a stepfile to
track the changes at each step - this is unlike the x86 ports where the
steps were only backpatched, so I just had to loop over all steps.

The idea with the stepfile is that for each iteration, the step is
increased as the very first change of the commit and all the build/test
issues are resolved until "make" runs successfully.

The pre-existing/older techdebt that genrom is almost an exact copy of
the version from x86-as is carried over for now.

genprims also still carries old baggage that I'll clean up later.

Rather than use 8 bit offsets just like in the x86 port, I've used 16 bit
offsets since the simh/PDP11 emulation provides the luxury of 64K of RAM.

An experiment with using a newer format for code.prims is documented in
the JOURNAL entry for 1 Dec '22