Change the output device to LPT which can be redirected to a file
It turns out that the easy solution to redirect the output of "emit" to
a file is to use the LPT device for output which can be redirected to a
file by attach'ing it to a named file. So that's what I've chosen to do
for now. See the JOURNAL entry for 3 Dec '22 for details.