PDP11 Port : Step 2 : Implement the stack (and dup/drop)
Given the PDP11's flexible addressing modes, it is trivial to implement
a stack. I've used the next available register (r2) for the stack pointer.
So far the register mapping is: R0: IP, R1: TOS, R2: SP

I'll save the machine stack with native SP (R6) for the return stack.

I've changed the earlier test for this in rom.4th to be more strict and
verified it doesn't break anything in the existing x86 port.