Split step 10 into smaller chunks (starting on x86 first)
Just like the refactoring that was needed for the conditionals, it
makes sense to split the and(&), or(|) and xor(^) bit operator
implementations into granular chunks rather than club them all into
one single step. As with the conditionals, I'm verifying this step
on the x86 ports first before doing it on the PDP11 port.

This refactoring has been reflected in the pseudo code steps as well.

Yet another techdebt item that now shows up is that the steps in the
pseudo code are disconnected from the rom.4th test steps - at some point
these need to be brought back into one coherent set of steps.