PDP11 Port : Step 10.3 : Implement the xor(|) bit operator
Trying to build this code ran into some trouble with the assembler.

The chapter on "Instruction Set for the PDP-11" in the book that I've
been using during this project mentions that XOR is a supported instruction,
on the PDP11 but the assembler was failing with the message "Unsupported
instruction set extension".

So far, I've been using the option "-mkd11d' which compiles for the
11/04 PDP-11 machine model as per the help page for pdp11-aout-as.

I'd initially chosen this option since the simh docs mention that the
PDP11/04 supports 64K of memory which is the smallest possible memory
that simh emulates for the PDP11.

Unfortunately, it now looks like the "-mkd11d" option limits the instruction
set such that it does not support the "xor" instruction. Since I don't
particularly care about the hardware configuration, for now, I've decided to
drop this option from the build and use whatever default it is that the
assembler chooses when nothing is specified.

That helped to sneak through the build issue with the assembler and the test
appears to be passing as well.

Also fixed a bug in the comments of the rom.4th xor test, while I was there.