PDP11 Port : Step 11 : Implement 2drop
The genprims script was a little too generous in the forthlabel
format which it accepted for direct pass through as an assembly
labels. That is fixed now so that "2drop" doesn't get passed
through as a label to the assembler.

Note for the unwary: 2drop also uses the lack of a terminating
semicolon in its definition as the "fall through" mechanism.