Move the shift operators to a later step
PDP11 has the ASL/ASR and ROL/ROR instructions to shift bits, but they
can only shift one bit at a time. The Forth shift operators need to support
multiple bit shifts though. So that implementation needs loop support (or a
means of unrolling the loop). Rather than deal with that now, I've deferred
it to a later step where loops are enabled. This needed me to do some manual
step renumbering so I hope it won't happen too often.

I adjusted the steps in the pseudo/code to also reflect the changed ordering
of operations.

The ordering changes also needed changes to test.expected and test.inp so
those changes are reflected in this commit as well.

The makefile was missing a step (bug introduced by an earlier commit) so I
went ahead and fixed that as well.

Finally, while I was there, I also changed the recent code to use #assert's
instead of directly calling emit.