PDP11 Port : Step 19 : Implement rp@!
Note that I could have used an "xchg" macro to eliminate the code
duplication seen in the identical pattern of code in sp@! and rp@!,
but this duplication is just too small to be worth the additional
context switch overhead of using a macro when reading code.

Just like the x86 port, the space allocated to the return stack is
entirely too small at just 10 bytes, but the nice thing is that since
the code is shared between all the ports, the limits will be reported
by the x86 emulation so I can do this without any fear of a stack overflow.