Add offset size specific flag around exec parameter handling in rom.4th
While working on the PDP11 port of exec I realized that the code in
rom.4th is currently designed for a system which uses a 1 byte offset
while the PDP11 uses 2 byte offsets. So I've delimited that code within
a #{if marker for portability.

I'm not going to worry about other offset sizes for now, I guess I can burn
that bridge after I've crossed it. ;)