Add "step wise" definitions in defs.4th to match the test cases in rom.4th
This change is in preparation for supporting definitions in the PDP11
port. Previously all of the definitions were pulled in one go in defs.4th.
But this breaks the step wise addition and testing of functionality - for
example, when definitions in def.4th depend on functionality that exists
only at a later step. So for now, I've worked around this by adding "steps"
to defs.4th in addition to rom.4th