Generalize the SERIAL port to SERIALIN and SERIALOUT
While preparing to implement p@/p! on the PDP11, I realized that although
the x86 can use just a single port and use IN/OUT opcodes for input/output
on ports, this may not be true on other architectures. So, for portability,
I've split it out into SERIALIN and SERIALOUT.

While I was there, I fixed up the expected input/output characters as well
which have changed due to the addition of newer steps at 4.{1,2,3}