More cleanup of variable names : RRRR -> RSTKSZ
Since I started doing a cleanup, I might as well just clean up the rest of
the variables which had meaningless placeholder names. The RRRR variable
was meant to be a configuration variable to set the size of the return
stack. So it has been replaced with a more meaningful name. Naming is hard.