PDP11 Port : Step 29 : Implement third
Since a definition of "third" exists in defs.4th, I didn't need to code
anything for this step, just test it. If this turns out to be a performance
sensitive operator, I'll need to figure out a way to override the Forth
definition with a native implementation.

$TECHDEBT
Need to figure out a means of overriding the definitions in defs.4th
with primitives defined in code.prims