Speed up the build after "make clean"
Currently after a "make clean", running make will result in a full
regression test of all previous steps (due to the "allsteps" target)

Since running the full regression test can take a while, I'll make it
a manual step rather than the default. This does make PDP11 the "primary"
target because the x86 build now depends on the PDP11 stepfile. I'll
revisit this decision if it turns out to be major issue down the line.