Discuss how much functionality needs to be shoehorned in ROM vs RAM
Microcontrollers come in various combinations of sizes of ROM and RAM.
Each of those variations imposes different constraints on the system
designer so a "one size fits all" approach to porting any given Forth
implementation will not cater to all of the differing size constraint
requirements that are sure to pop up out in the field. This JOURNAL
entry is just an attempt at documenting my current thinking on how to
cater to all these different constituencies.