Add dictionary headers for all of the PDP11 "words"
This set of changes, primarily to the genprims and genrom scripts on
PDP11, is to prefix each of the "words" generated from code.prims and
defs.4th by these scripts with a "dictionary header"

The dictionary header follows the Forth convention of encapsulating all of
of the usual fields viz: link(lfa), name(nfa) and code(cfa) which could be
described using the C-like pseudo code notation below:
	struct counted_string {
		byte count;
		byte name[count];
	};
	struct dict {
		struct dict *lfa; // links to the previous dictionary entry
		struct counted_string nfa;
		byte cfa[...];    // the "code field" extends out as needed
	}

The addition of the headers currently results in a net addition of 542 bytes
based on the last offsets listed in the .lst files (0x840 vs 0xa5e). As a
percentage, this works out to ~26% (=2654/2112) so there is some value in
not having a dictionary around for the ROM/RAM constrained processors.

Since we need to cater to a wide range of microprocessors, which may
be constrained in both RAM and ROM (see the earlier JOURNAL entry dated
20 Dec '22), it makes sense to split out the functionality into two separate
binaries. The new binary "forth_dict", which contains the dictionary can be
tested and/or flashed independently of the umbilically hosted binary (which
continues to be called "forth"). Most of the makefile changes are to support
this split.

I've documented, in gory/minute detail, the debug steps I took while hunting
down a bug that I ran into, while making these changes, in the JOURNAL entry
dated 22 Dec '22 just as a reminder of the debug steps that were required in
case more such "debug opportunities" show up in the course of future ports.

The way that the "latest" variable's content is generated is rather
convoluted, but I don't see a simpler approach for now. $TECHDEBT?

There is currently no test to go with these changes to verify that the
header entries are sane and generated correctly. For now I assume that a
manual/visual inspection suffices until the rest of the code is in place
but before I do that I want to repeat this step for the x86 ports as well.