Fix the PDP11 build (and try to unify all the makefiles)
The PDP11 build was failing due to a dependency issue when running
"make allsteps". For now I've fixed it by adding "step ifdefs" to the
code.prims in addition to defs.4th and rom.4th

Rather than have it be a one-off just for PDP11 I then made the same
change to the x86 ports as well so that most of the makefile continue
to share a common template. I'm tempted to over-engineer these makefiles
for using a common template and adding abstractions but I think I'll
successfully resist that tempation.

I'm sure there's a better fix for this but for now, this duct tape
should be more than sufficient to hold things together.