Implement "same" to check if two strings are the same or not
Now that the dictionary is setup in ROM (to my satisfaction), I can start to
add in the definitions to look up words in the dictionary. The first such
definition that I'll add is "same" which checks for string equality. It
works just like memcmp except that the return values are reversed.

I've also introduced "cell" which is just a wrapper over the arch dependent
constant for how wide a "dw/.word" directive in assembly expands to. It is
needed now since it is required to step over the link field.