Implement "match" to compare a counted string, given an lfa
While the previous commit implemented "same" which just does a string
comparison akin to C's strcmp/memcmp, "match" wraps over it to provide
additional functionality specific to Forth by checking the count byte
(since the dictionary names are counted strings) and also stepping over
the link field of the dictionary entry that was passed on the stack.

It can also find words that were marked "immediate", which, following Forth
convention, sets the MSB bit in the count byte in the name field. Since the
current set of dictionary entries does not have any "immediate" words, test
coverage for that part of the definition is deferred until that is added.