Implement the allocation routines "alloc" and "alloca" (alloc aligned)
In preparation for being able to add dictionary entries, we need allocation
routines, so that is provided as part of this step. It is debatable whether
these definitions should be provided at an earlier step but my reasoning is
that fourforth is the only version which assumes anything other than tiny
amounts of RAM. So, an allocator at any of the earlier versions might just
be overkill.

Since we are now finally moving on to a fully interactive version of Forth,
the default makefile target of fourforth has also been added although there
is no difference in terms of the build itself.

I'd forgotten to bump up the step in the previous commit so the step jumps
up by 2 in the makefile and by 1 in the stepfile as part of this commit.