Implement ':' and ';'
By leveraging the test code that was used in the previous step, it is fairly
easy to have working and tested code as part of this implemention. Despite
the previous tests, there are some new tests at this step as well just as a
sanity check. Since they are being tested "out of band" in some sense, at
this step, ';' does not need to be an "immediate" word, yet.