Rename compdef to ESCAPE
Naming really is hard. I've decided that I want to use compdef to denote
"compile a definition" (ie append the cfa of into the dictionary). The
current "compdef" appears to be misnamed and I think it should really be
called some variation of "escape". Looking at older implementations,
"escape" might have been called COMPILE (or is it [COMPILE]?) in those
implementations but I don't want to add to the overall confusion in case
this is subtly different from that. So I'm just going to call it ESCAPE.
The all caps is meant to call attention to the fact that it cannot be
wrapped up in other definitions.