Implement the "outer" interpreter
So far the repl has been missing its 'l' - the looping part. This was
because I needed a means of exiting back to the test code to verify the
test results. Now that things are mostly working, I can start using the
repl in earnest by wrapping it in a loop but I still need a means of
something like C's "longjmp" and for this I've implemented "3ret" which
is meant to be called within the "outer" repl to exit back to its caller.

Rather than append the rest of the code to test.inp which was meant only
to contain input for the various tests, I've decided to stick the rest of
the code to be parse'd by the repl into rest.inp. Since I'm adding another
file, some of the makefile dependencies needed to be updated as well.