Implement conditionals: if{ ... }if
To keep things simple, I'm not implementing "}else{" at this point and will
defer that to the next commit. Since I need a means of testing the new code,
"?dup" was implemented as part of writing a test for the new "if{ .. }if"
conditional control structure.

Since rest.inp is now turning into a mix of code and test, I've decided to
rearrange the existing code a bit as well so as to wrap the code that is
required only for the tests under "ifdef TESTROM" and leave the rest of the
new code within the "#if step" wrappers (outside the "#ifdef TESTROM").
Hopefully this makes it easier to distinguish the "resident" code from the
"test only" code.