Extend conditionals to support "}else{"
Finishing up the support for "if{ ... }else{ ... }if" ended up needing an
implementation of "[compile]" as well. Rather than have "[compile]" read in
an immediate word, and compile it in, effectively overriding the logic of
the outer interpreter, which seems to be the usual implementation used in
a couple of Forth implementations that I looked at, I've decided instead to
continue to keep the "exec vs compile" decision centralized in the outer
interpreter. In this new scheme, "[compile]" only sets an "override" flag
(which has been shoehorned into the state flag). This flag is used in the
modified outer interpreter to handle the exec vs compile decision.

Rather than have to run "make allsteps" since I'm changing old code, I've
put the modified code with the "#if step" using the new step value, but I
did run a regression test anyway just to make sure nothing broke.