Implement '.' for debugging
I've taken liberties with the implementation of '.' to make it simpler and
so it currently prints the top of stack in hex and embedded in spaces with
a prefix of 0x so there is no ambiguity at all about what the base is.

While I was there, since I prefer "def{" and "}def" to the conventional
Forth defining words (':' and ';'), I've snuck them in as well, as part of
this commit. This will be the last set of feature additions to the repl
though. The JOURNAL entry for 26 Jan '23 goes into some detail regarding my
reasoning about why it makes sense to make this a good stopping point.

All additional changes from here on out will be for supporting newer ports
to other architectures/microcontrollers/cpus.