C Port : step 0 : A halting interpreter
Restart the C port from scratch

The JOURNAL entry goes into more detail but this is just a fresh start
of the C port which had been put on indefinite hiatus earlier.

Most of the scripts are just minor tweaks of the PDP11 equivalents.
Additional changes will be made as and when necessary.

This is adding on still more $TECHDEBT because of even more copy/paste of
the Perl scripts. At some point I should try to unify the genrom and
genprims scripts across ports.