C Port : step 5 : Implement support for control stuctures
The control structures implemented in this step are:
	if{ ... }if
	if{ ... }else{ ... }if
	loop{ ... }while{ ... }loop and
	loop{ ... }until{ ... }loop

While assembly code can compute offsets easily, C does not allow that so
all of the jump offsets are calculated and filled into the rom code in a
second pass implemented by the patchrom Perl script.

See the comments in the patchrom code for details.