C Port : step 18 : Implement sp@!
In the assembly version of this code, I could swap pointers since I could
sling any value between registers, but here, I need to be more circumspect.

Rather than swap pointers, I've weaseled out by using an implementation
that uses a multi dimensional array (of stacks) and sp@! just switches
between them. The initial value is set to stack[1] since the test switches
it to stack[0] and back, although in theory, there should be no issues with
switching to the same stack that you are already on.

The 0'th element of each stack contains the offset to the top of the stack