C Port : step 19 : Implement rp@!
Ideally we should just ignore the rp@! initialization done in rom.4th
since it was inherited from the PDP11 port and it applies mostly to a
microcontroller world where the entire memory is under software control.

What I'm trying to do here is to try and shoehorn that model on top of
the C runtime running on top of an OS so that the testcode doesn't need
to be changed. Hopefully, all this additional work is worth it.

If the contents of "here" gets pushed out (due to introduction of even
more variables) such that its content is larger than nrstacks, there will
be corruption caused by writing outside the array boundary so that's
something to watch out for - I'll either bump up nrstacks or think of a
workaround if we ever run into that situation.