C Port : step 23 : Implement call
Since the C implementation has a "machine" stack which is distinct from the
"return" stack, implementing "call" required some contortions which included
adding yet another configuration parameter which I've called "machinestack"
which can be set to 0 or 1. The x86 and PDP11 configs needed to be modified
to use 0 for this setting while the C implementation uses 1. Since common
code was modified, make allsteps was run to verify no regressions crept in.