C Port : step 26 : Verify "until" loops (also, modified p@)
Since the test code expects p@ to work in a specific way on a specific
port and this is inherited from the PDP11 code, it doesn't cost anything
to keep the "API", if you want to call it that, the same on the C port
as well.

The "until" loop functionality itself is generic, so no other code
changes specific to this port is required, other than to verify that it
works as expected.