C Port : step 32 : Modify i to handle machinestack config
Since the C port uses a "machine" stack which is independent from the return
stack, routines that previously assumed a particular configuration of the
return stack will need modification. "i" is the first such definition that
needs to be modified.

Since this commit modified common code (in defs.4th), a full regression test
was done by running "make allsteps"