C Port : step 46 : Verify definitions can be run from the repl
The earlier set of changes at steps 44 and 45 have made this step also
diverge from the older implementation. I need to go back and make this
uniform across the different ports so I'll mark this as $TECHDEBT for now.

Since common code was modified, so as usual, I ran the full set of
regression tests (by running "make allsteps").