C Port : step 52 : Verify that definitions using literals work in the repl
This step ran into a segv due to a return stack overflow - more details are
in the JOURNAL entry for 01 Mar '23. I'll note a $TECHDEBT entry here to
track a better means for sizing the return stack appropriately

Since there were changes to common code and mods to the config files for
the earlier architectures, a full regression using "make allsteps" was done.