Use the 'runallsteps' script across all ports to "make allsteps"
The "allsteps" regression test, for the x86 and PDP11 ports was simple
enough to be coded directly in the makefile itself. For the C port though,
the 'runallsteps' Perl script was added since it needed to do stuff which
would be difficult to describe within the confines of a makefile.

Rather than have this logic in two places, I decided to unify it by moving
the code from the makefile into the script.