x86-user : all in on 64-bits
Although my development system uses 64 bits I've tried to limit the size of
literals to 32 bits, but the previous commit shows that this can be a source
of subtle bugs. So I've decided to use 64 bits throughout the code since it
simplifies things quite a bit.

Since I needed to run the regression only for the x86-user port, I decided
to also extend the runallsteps script so it can be run for just one port
instead of all ports (since the PDP11 emulation is especially slow).