x86-user : step 61 : Enable for{ ... }for loop support in the repl
While implementing this step, I realized that "over" needs to be a primitive
(well, it can be a definition, but that would make the definition of
"compile" more complicated than it already is, and would need even more
parameterization). So for now, I've implemented it as a primitive but since
I was tight on space in the 256 byte area, I moved it to the end so it can
overhang past the 256 byte boundary (since only the starting byte needs to
be within the 256 byte boundary).