x86-user : step 62 : Implement '.' to conclude the x86 assembly user port
The x86 userland port (in assembly) has been fairly smooth sailing, so far,
until I ran into trouble at this step, at the very end. The symptom as usual
was a SIGSEGV which was surprising since all this step does is add the
definition for '.' and test that it works as expected. So after feeling
pretty puzzled for a while, I started looking at debug options.

The first sanity check I did was to move the definition from rest.inp to
defs.4th and verify that there were no issues with the definition itself.
From that, it was clear that the problem was in the code generation in
the outer interpreter. So the next debug step was to add the same definition
of '.' in both rest.inp and defs.4th and then compare the bytes. To do this,
I then needed to add a debug word called "showcfa" which takes the cfa and
dumps bytes until it sees an 'exit' byte code.

Luckily, despite the diffs between the two, I was able to filter it down to
see that the offset generated for the jump at the end of the while loop was
a positive value (20) instead of a negative value (the defs_dict.s generated
from the code in defs.4th showed the expected negative offset of 0x9A).

I then realized that the offset was positive because the code had become
too large and overflowed the implicit expectation that the body of a loop
(or a conditional) cannot exceed 255 bytes since the offset is stored using
just one byte. So I had to split the body of the definition to call "nybble"
and this fixed the bug.