x86-sys : step 60 : Enable loop{ ... }until{ ... }loop in the repl
This was a difficult bug to figure out. I guess I got lucky because I
noticed that the address on which the memory access was faulting was
ending in a couple of zeroes. So on a hunch I decided to bump up the
memory allocated since there is nothing else that is substantially
different between this code and the code in x86-user.