m68k : step 27 : Verify "while" loops
Whew! that was a long and drawn out debug session. It turned out to be a
latent bug in the implementation of all of the branch operators (j/jz/jnz).

The code for the branch operators was changed from the usual template
since the `move` opcode in the m68k changes the flags. So although the
code took care to `drop` the top of stack in the branching part, I forgot
to do it in the straight line path.

I'll add a $TECHDEBT here since I think some additional tests to verify
the branch code at a much earlier step would have been definitely better.