Update the PORTING guide
The "guide" itself is too high level to be useful now but is a good start.

Since there are so many x86 variants, I tried to create a table that
summarizes the tiny differences in detail between each of them and also
captures the information of all the ports that have been completed.