z80 : step 3 : Implement literals
The easier implementation for lit might be to use `ld bc, (ix)` but that
appears to be an ez80 only instruction so I've used the equivalent expanded
version which works without making the z80 -> ez80 transition.

I'm now thinking it would be much easier to just code the entire thing in C
and let SDCC generate the code for Z80 and then pick and choose from it for
the actual assembly version. Just like x86, that will give me a means of
checking how good SDCC is at generating compact code compared to this manual
grunt work.

It obviously also has the nice advantage of getting ports "for free" to all
of the CPUs which are supported by sdcc.