z80-c : step 0 : A halting interpreter, exit only
As I mentioned in the previous commit message, I'm going ahead with
an implementation for Z80 in C/sdcc. Unlike gcc, sdcc does not support
computed goto's so I can't use the earlier C version used for x86 as is,
so, this version will be using a plain old switch style interpreter.