z80-c : step 4.2 : Implement jump if zero (jz)
I spent a fair bit of time debugging what appeared to be a bug in sdcc
(or atleast a difference in the way that sdcc and gcc handle sign extension
from a char to an int). It turns out that on the sdcc version that I'm using
char is unsigned by default which I think violated the law of least surprise
at least for me, but after almost a day and a half of laborious debugging,
the workaround I figured out was to just declare a "signed char". Grrr!