z80-c : step 26 : Verify "until" loops
When initializing a char array with a signed value, sdcc 4.2.0 whines saying
	warning 158: overflow in implicit constant conversion
This happens during the initialization of the rom array and the value could
be negative for backward jumps. Rather than modify any code, I've decided to
workaround this issue in the `patchrom` script so that it just generates an
unsigned 2's complement byte if you pass it the `--sdcc` option.