z80-c : step 52 : Verify that definitions using literals work in the repl
Yet another configuration variable needed to be added to fix this on z80-c

The new configuration parameter is called 'PRIMSZ' and it denotes the
number of bytes required to point to the start address/enum value of the
definition of the primitive.

I've added this value to the config files of all of the ports and modified
rom.4th to handle this configuration appropriately.

Since common code was touched, a full regression test using `runallsteps`
was done.