z80-c : step 62 : Implement '.' to conclude the x86 assembly syscall port
Just like the x86 userland port, I got caught in a long and drawn out
debugging marathon at this very last step of the sdcc/z80 port.

Since this is in C, my knee jerk reaction was to obviously reach for the
heavy guns that I'd already implemented for the gcc/x86_64 port a while back

Unfortunately, that resulted in taking a detour to figure out how `printf`
could be layered over the `putchar` implementation which had been used to
implement `emit`.

With that out of the way, the next problem I had to deal with was that the
`printf` debugging was too verbose and excruciatingly slow. To alleviate the
slowness, I added a debug flag which was accessible from both C and Forth to
filter down the logging to a manageable level and this helped me track the
problem down to jump offsets not being generated correctly. Reviewing my
changes to the `patch` function in `rest.inp` led me to the root cause that
I'd left a `dec` in the code which was needed only for the 16-bit offset
not in the current code which only uses an 8-bit offset. Special casing it
out fixed that issue and luckily all of the other changes held their ground.