z80 : fix a stack overflow caused by a bug in c!
c! was not pop'ping an element from the stack and this resulted in a crash
observed at step 39. So I've fixed the implementations of c! and c@ to be
"more correct". Regression testing with the fix shows that the last step
that successfully passes now has regressed to step 25, so the stepfile was
reverted all the way back to that step so that each additional fix that
needs to be made from here on can have it's own commit rather than get
lumped into one big fix.