z80 : Fix bugs in the left and right shift ops
I may have misunderstood the diagrams shown in the Z80 Assembly Language
Programming book by Lance A Levental for the sla and srl opcodes which
show the shift happening through the high byte from the low byte (and
vice versa). Rather than check to see if this is just an emulator bug,
I've changed the code so that it should always work. This allows the next
step 40 to pass since atoi which is used at step 40 uses the shift ops.

In any case, I've extended the tests for the shift ops so that bugs such as
these can be caught earlier.